fix: cspell-tools - compounds
This PR changes a single rule when merging compound forms:

When a word is allowed to be stand alone and a midfix compound, then allow it to be a prefix as well as a suffix.

The logic:
If we allow `code` and `strongcodetype`, then `codetype` or `strongcode` should also be allowed.